split SAM economics test into separate test_economics_sam file
