+++ tests/test_geophires_x.py
@@ -913,29 +913,3 @@ Print Output to Console, 1"""
             )
             client.get_geophires_result(params)
         self.assertIn('Electricity production calculated as negative', str(e.exception))
-
-    def test_economic_model_single_owner_ppa_sam(self):
-        def _get_result(_params) -> GeophiresXResult:
-            return GeophiresXClient().get_geophires_result(
-                GeophiresInputParameters(
-                    from_file_path=self._get_test_file_path('geophires_x_tests/generic-egs-case.txt'),
-                    params={'Economic Model': 5, **_params},
-                )
-            )
-
-        def _lcoe(r: GeophiresXResult) -> float:
-            return r.result['SUMMARY OF RESULTS']['Electricity breakeven price']['value']
-
-        def _npv(r: GeophiresXResult) -> float:
-            return r.result['ECONOMIC PARAMETERS']['Project NPV']['value']
-
-        base_result = _get_result({})
-        base_lcoe = _lcoe(base_result)
-        self.assertGreater(base_lcoe, 6)
-
-        npvs = [_npv(_get_result({'Starting Electricity Sale Price': x / 100.0})) for x in range(1, 20, 2)]
-        for i in range(len(npvs) - 1):
-            self.assertLess(npvs[i], npvs[i + 1])
-
-        with self.assertRaises(RuntimeError):
-            _get_result({'End-Use Option': 2})
